+++ libTidyEngine/mesh.hpp
@@ -27,7 +27,7 @@ Contact the author at: jakob.sinclair99@gmail.com
 class Mesh: public Renderable {
 public:
 	Mesh();
-	Mesh(std::vector<Vertex> &vertices, std::vector<uint64_t> &indices, std::vector<GLuint> &textures);
+	Mesh(std::vector<Vertex> &vertices, std::vector<uint32_t> &indices, std::vector<GLuint> &textures);
 	virtual ~Mesh();
 
 	virtual void Render();
